+++ platformio.ini
@@ -108,7 +108,6 @@ build_flags =
 # Used to run unit test; g++ must be in PATH.
 platform = native
 build_flags = "-std=c++11"
-lib_ldf_mode = chain+
 lib_deps =
     ArduinoFake
 	ArduinoJson@^6.21.0
