+++ src/Config/config.hpp
@@ -4,8 +4,8 @@
 #include <stdint.h>
 #include <LittleFS.h>
 #include <ArduinoJson.h>
-#include <canBus.hpp>
-#include <x10a.hpp>
+#include "CAN/canBus.hpp"
+#include "X10A/x10a.hpp"
 
 #ifdef ARDUINO_ARCH_ESP8266
 #define FILE_READ       "r"
