+++ src/X10A/arrayFunctions.cpp
