+++ src/X10A/arrayFunctions.hpp
