+++ src/X10A/comm.cpp
