+++ src/X10A/registryBuffer.hpp
