+++ test/test_rotex_protocol.cpp
@@ -4,7 +4,23 @@
 #include <fstream>
 #include <ArduinoJson.h>
 
-#include "x10a.hpp"
+class PrintF
+{
+public:
+    PrintF() {}
+    virtual ~PrintF() {}
+
+    size_t printf(const char * format, ...) {return 0;}
+    size_t print(const char[]) {return 0;}
+};
+
+#define SERIAL_TYPE PrintF
+#define SERIAL_CONFIG (SWSERIAL_8E1)
+#define DEFINE_SerialX10A SERIAL_TYPE SerialX10A
+typedef unsigned long ulong;
+
+#include "X10A/x10a.hpp"
+#include "Config/config.hpp"
 
 #define MODELS_CONFIG_SIZE 1024*10
 
@@ -15,6 +31,7 @@
 #endif
 
 X10A_Config* X10AConfig = nullptr;
+
 
 void setUp(void)
 {
